create classes for generators
