+++ pandapower/converter/cim/cim2pp/build_pp_net.py
@@ -44,207 +44,6 @@ class CimConverter:
             df[sc['o_cl']] = cim_type
         return df
 
-    def _convert_synchronous_machines_cim16(self):
-        time_start = time.time()
-        self.logger.info("Start converting SynchronousMachines.")
-        eqssh_synchronous_machines = self._prepare_synchronous_machines_cim16()
-
-        # convert the SynchronousMachines with voltage control to gens
-        eqssh_sm_gens = eqssh_synchronous_machines.loc[(eqssh_synchronous_machines['mode'] == 'voltage') &
-                                                       (eqssh_synchronous_machines['enabled'])]
-        self.copy_to_pp('gen', eqssh_sm_gens)
-        # now deal with the pq generators
-        eqssh_synchronous_machines = eqssh_synchronous_machines.loc[(eqssh_synchronous_machines['mode'] != 'voltage') |
-                                                                    (~eqssh_synchronous_machines['enabled'])]
-        self.copy_to_pp('sgen', eqssh_synchronous_machines)
-        self.logger.info("Created %s gens and %s sgens in %ss." %
-                         (eqssh_sm_gens.index.size, eqssh_synchronous_machines.index.size, time.time() - time_start))
-        self.report_container.add_log(Report(
-            level=LogLevel.INFO, code=ReportCode.INFO_CONVERTING,
-            message="Created %s gens and %s sgens from SynchronousMachines in %ss." %
-                         (eqssh_sm_gens.index.size, eqssh_synchronous_machines.index.size, time.time() - time_start)))
-
-    def _prepare_synchronous_machines_cim16(self) -> pd.DataFrame:
-        eq_generating_units = self.cim['eq']['GeneratingUnit'][['rdfId', 'nominalP', 'minOperatingP', 'maxOperatingP']]
-        # a column for the type of the static generator in pandapower
-        eq_generating_units['type'] = 'GeneratingUnit'
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['WindGeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('WP', inplace=True)
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['HydroGeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('Hydro', inplace=True)
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['SolarGeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('PV', inplace=True)
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['ThermalGeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('Thermal', inplace=True)
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['NuclearGeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('Nuclear', inplace=True)
-        eq_generating_units.rename(columns={'rdfId': 'GeneratingUnit'}, inplace=True)
-        eqssh_synchronous_machines = self.merge_eq_ssh_profile('SynchronousMachine', add_cim_type_column=True)
-        if 'type' in eqssh_synchronous_machines.columns:
-            eqssh_synchronous_machines.drop(columns=['type'], inplace=True)
-        if 'EquipmentContainer' in eqssh_synchronous_machines.columns:
-            eqssh_synchronous_machines.drop(columns=['EquipmentContainer'], inplace=True)
-        eqssh_synchronous_machines = pd.merge(eqssh_synchronous_machines, eq_generating_units,
-                                              how='left', on='GeneratingUnit')
-        eqssh_reg_control = self.merge_eq_ssh_profile('RegulatingControl')[['rdfId', 'mode', 'enabled', 'targetValue', 'Terminal']]
-        # if voltage is not on connectivity node, topological node will be used
-        eqtp_terminals = pd.merge(self.cim['eq']['Terminal'], self.cim['tp']['Terminal'], how='left', on='rdfId')
-        eqtp_terminals.ConnectivityNode.fillna(eqtp_terminals.TopologicalNode, inplace=True)
-        eqtp_terminals = eqtp_terminals[['rdfId', 'ConnectivityNode']].rename(columns={'rdfId': 'Terminal', 'ConnectivityNode': 'reg_control_cnode'})
-        eqssh_reg_control = pd.merge(eqssh_reg_control, eqtp_terminals, how='left', on='Terminal')
-        eqssh_reg_control.drop(columns=['Terminal'], inplace=True)
-        # add the voltage from the bus
-        eqssh_reg_control = pd.merge(eqssh_reg_control, self.net.bus[['vn_kv', sc['o_id']]].rename(columns={sc['o_id']: 'reg_control_cnode'}), how='left', on='reg_control_cnode')
-        # merge with RegulatingControl to check if it is a voltage controlled generator
-        eqssh_reg_control = eqssh_reg_control.loc[eqssh_reg_control['mode'] == 'voltage']
-        eqssh_synchronous_machines = pd.merge(
-            eqssh_synchronous_machines, eqssh_reg_control.rename(columns={'rdfId': 'RegulatingControl'}),
-            how='left', on='RegulatingControl')
-        eqssh_synchronous_machines = pd.merge(eqssh_synchronous_machines, self.bus_merge, how='left', on='rdfId')
-        eqssh_synchronous_machines.drop_duplicates(['rdfId'], keep='first', inplace=True)
-        eqssh_synchronous_machines['vm_pu'] = eqssh_synchronous_machines.targetValue / eqssh_synchronous_machines.vn_kv
-        eqssh_synchronous_machines['vm_pu'].fillna(1., inplace=True)
-        eqssh_synchronous_machines.rename(columns={'vn_kv': 'bus_voltage'}, inplace=True)
-        eqssh_synchronous_machines['slack'] = False
-        # set the slack = True for gens with highest prio
-        # get the highest prio from SynchronousMachines
-        sync_ref_prio_min = eqssh_synchronous_machines.loc[
-            (eqssh_synchronous_machines['referencePriority'] > 0) & (eqssh_synchronous_machines['enabled']),
-            'referencePriority'].min()
-        # get the highest prio from ExternalNetworkInjection and check if the slack is an ExternalNetworkInjection
-        enis = self.merge_eq_ssh_profile('ExternalNetworkInjection')
-        regulation_controllers = self.merge_eq_ssh_profile('RegulatingControl')
-        regulation_controllers = regulation_controllers.loc[regulation_controllers['mode'] == 'voltage']
-        regulation_controllers = regulation_controllers[['rdfId', 'targetValue', 'enabled']]
-        regulation_controllers.rename(columns={'rdfId': 'RegulatingControl'}, inplace=True)
-        enis = pd.merge(enis, regulation_controllers, how='left', on='RegulatingControl')
-
-        eni_ref_prio_min = enis.loc[(enis['referencePriority'] > 0) & (enis['enabled']), 'referencePriority'].min()
-        if pd.isna(sync_ref_prio_min):
-            ref_prio_min = eni_ref_prio_min
-        elif pd.isna(eni_ref_prio_min):
-            ref_prio_min = sync_ref_prio_min
-        else:
-            ref_prio_min = min(eni_ref_prio_min, sync_ref_prio_min)
-
-        eqssh_synchronous_machines.loc[eqssh_synchronous_machines['referencePriority'] == ref_prio_min, 'slack'] = True
-        eqssh_synchronous_machines['p_mw'] = -eqssh_synchronous_machines['p']
-        eqssh_synchronous_machines['q_mvar'] = -eqssh_synchronous_machines['q']
-        eqssh_synchronous_machines['current_source'] = True
-        eqssh_synchronous_machines['sn_mva'] = \
-            eqssh_synchronous_machines['ratedS'].fillna(eqssh_synchronous_machines['nominalP'])
-        # SC data
-        eqssh_synchronous_machines['vn_kv'] = eqssh_synchronous_machines['ratedU'][:]
-        eqssh_synchronous_machines['rdss_ohm'] = \
-            eqssh_synchronous_machines['r2'] * \
-            (eqssh_synchronous_machines['ratedU']**2 / eqssh_synchronous_machines['ratedS'])
-        eqssh_synchronous_machines['xdss_pu'] = eqssh_synchronous_machines['x2'][:]
-        eqssh_synchronous_machines['voltageRegulationRange'].fillna(0., inplace=True)
-        eqssh_synchronous_machines['pg_percent'] = eqssh_synchronous_machines['voltageRegulationRange']
-        eqssh_synchronous_machines['k'] = (eqssh_synchronous_machines['ratedS'] * 1e3 / eqssh_synchronous_machines[
-            'ratedU']) / (eqssh_synchronous_machines['ratedU'] / (
-                    3 ** .5 * (eqssh_synchronous_machines['r2'] ** 2 + eqssh_synchronous_machines['x2'] ** 2)))
-        eqssh_synchronous_machines['rx'] = eqssh_synchronous_machines['r2'] / eqssh_synchronous_machines['x2']
-        eqssh_synchronous_machines['scaling'] = 1.
-        eqssh_synchronous_machines['generator_type'] = 'current_source'
-        eqssh_synchronous_machines.rename(columns={'rdfId_Terminal': sc['t'], 'rdfId': sc['o_id'],
-                                                   'connected': 'in_service', 'index_bus': 'bus',
-                                                   'minOperatingP': 'min_p_mw', 'maxOperatingP': 'max_p_mw',
-                                                   'minQ': 'min_q_mvar', 'maxQ': 'max_q_mvar',
-                                                   'ratedPowerFactor': 'cos_phi'}, inplace=True)
-        return eqssh_synchronous_machines
-
-    def _convert_asynchronous_machines_cim16(self):
-        time_start = time.time()
-        self.logger.info("Start converting AsynchronousMachines.")
-        eqssh_asynchronous_machines = self._prepare_asynchronous_machines_cim16()
-        self.copy_to_pp('motor', eqssh_asynchronous_machines)
-        self.logger.info("Created %s motors in %ss." %
-                         (eqssh_asynchronous_machines.index.size, time.time() - time_start))
-        self.report_container.add_log(Report(
-            level=LogLevel.INFO, code=ReportCode.INFO_CONVERTING,
-            message="Created %s motors from AsynchronousMachines in %ss." %
-                         (eqssh_asynchronous_machines.index.size, time.time() - time_start)))
-
-    def _prepare_asynchronous_machines_cim16(self) -> pd.DataFrame:
-        eq_generating_units = self.cim['eq']['WindGeneratingUnit'].copy()
-        # a column for the type of the static generator in pandapower
-        eq_generating_units['type'] = 'WP'
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['GeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('GeneratingUnit', inplace=True)
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['HydroGeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('Hydro', inplace=True)
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['SolarGeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('PV', inplace=True)
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['ThermalGeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('Thermal', inplace=True)
-        eq_generating_units = pd.concat([eq_generating_units, self.cim['eq']['NuclearGeneratingUnit']], sort=False)
-        eq_generating_units['type'].fillna('Nuclear', inplace=True)
-        eq_generating_units.rename(columns={'rdfId': 'GeneratingUnit'}, inplace=True)
-        eqssh_asynchronous_machines = self.merge_eq_ssh_profile('AsynchronousMachine', add_cim_type_column=True)
-        # prevent conflict of merging two dataframes each containing column 'name'
-        eq_generating_units.drop('name', axis=1, inplace=True)
-        eqssh_asynchronous_machines = pd.merge(eqssh_asynchronous_machines, eq_generating_units,
-                                               how='left', on='GeneratingUnit')
-        eqssh_asynchronous_machines = pd.merge(eqssh_asynchronous_machines, self.bus_merge, how='left', on='rdfId')
-        eqssh_asynchronous_machines['p_mw'] = -eqssh_asynchronous_machines['p']
-        eqssh_asynchronous_machines['q_mvar'] = -eqssh_asynchronous_machines['q']
-        eqssh_asynchronous_machines['current_source'] = True
-        eqssh_asynchronous_machines['cos_phi_n'] = eqssh_asynchronous_machines['ratedPowerFactor'][:]
-        eqssh_asynchronous_machines['sn_mva'] = \
-            eqssh_asynchronous_machines['ratedS'].fillna(eqssh_asynchronous_machines['nominalP'])
-        eqssh_asynchronous_machines['generator_type'] = 'async'
-        eqssh_asynchronous_machines['loading_percent'] = \
-            100 * eqssh_asynchronous_machines['p_mw'] / eqssh_asynchronous_machines['ratedMechanicalPower']
-        eqssh_asynchronous_machines.rename(columns={'rdfId_Terminal': sc['t'], 'rdfId': sc['o_id'],
-                                                    'connected': 'in_service', 'index_bus': 'bus',
-                                                    'rxLockedRotorRatio': 'rx', 'iaIrRatio': 'lrc_pu',
-                                                    'ratedPowerFactor': 'cos_phi', 'ratedU': 'vn_kv',
-                                                    'efficiency': 'efficiency_n_percent',
-                                                    'ratedMechanicalPower': 'pn_mech_mw'}, inplace=True)
-        eqssh_asynchronous_machines['scaling'] = 1
-        eqssh_asynchronous_machines['efficiency_percent'] = 100
-        return eqssh_asynchronous_machines
-
-    def _convert_energy_sources_cim16(self):
-        time_start = time.time()
-        self.logger.info("Start converting EnergySources.")
-        eqssh_energy_sources = self._prepare_energy_sources_cim16()
-        es_slack = eqssh_energy_sources.loc[eqssh_energy_sources.vm_pu.notna()]
-        es_sgen = eqssh_energy_sources.loc[eqssh_energy_sources.vm_pu.isna()]
-        self.copy_to_pp('ext_grid', es_slack)
-        self.copy_to_pp('sgen', es_sgen)
-        # self._copy_to_pp('sgen', eqssh_energy_sources)
-        self.logger.info("Created %s sgens in %ss." % (eqssh_energy_sources.index.size, time.time() - time_start))
-        self.report_container.add_log(Report(
-            level=LogLevel.INFO, code=ReportCode.INFO_CONVERTING,
-            message="Created %s sgens from EnergySources in %ss." %
-                    (eqssh_energy_sources.index.size, time.time() - time_start)))
-
-    def _prepare_energy_sources_cim16(self) -> pd.DataFrame:
-        eq_energy_scheduling_type = \
-            pd.concat([self.cim['eq']['EnergySchedulingType'], self.cim['eq_bd']['EnergySchedulingType']], sort=False)
-        eq_energy_scheduling_type.rename(columns={'rdfId': 'EnergySchedulingType', 'name': 'type'}, inplace=True)
-        eqssh_energy_sources = self.merge_eq_ssh_profile('EnergySource', add_cim_type_column=True)
-        eqssh_energy_sources = pd.merge(eqssh_energy_sources, eq_energy_scheduling_type, how='left',
-                                        on='EnergySchedulingType')
-        eqssh_energy_sources = pd.merge(eqssh_energy_sources, self.bus_merge, how='left', on='rdfId')
-        eqssh_energy_sources.drop_duplicates(['rdfId'], keep='first', inplace=True)
-        sgen_type = dict({'WP': 'WP', 'Wind': 'WP', 'PV': 'PV', 'SolarPV': 'PV', 'BioGas': 'BioGas',
-                          'OtherRES': 'OtherRES', 'CHP': 'CHP'})  # todo move?
-        eqssh_energy_sources['type'] = eqssh_energy_sources['type'].map(sgen_type)
-        eqssh_energy_sources['p_mw'] = -eqssh_energy_sources['activePower']
-        eqssh_energy_sources['q_mvar'] = -eqssh_energy_sources['reactivePower']
-        eqssh_energy_sources['va_degree'] = eqssh_energy_sources['voltageAngle'] * 180 / math.pi
-        eqssh_energy_sources['vm_pu'] = \
-            eqssh_energy_sources['voltageMagnitude'] / eqssh_energy_sources['base_voltage_bus']
-        eqssh_energy_sources['scaling'] = 1.
-        eqssh_energy_sources['current_source'] = True
-        eqssh_energy_sources['generator_type'] = 'current_source'
-        eqssh_energy_sources.rename(columns={'rdfId_Terminal': sc['t'], 'rdfId': sc['o_id'], 'connected': 'in_service',
-                                             'index_bus': 'bus'}, inplace=True)
-        return eqssh_energy_sources
-
     def _convert_static_var_compensator_cim16(self):
         time_start = time.time()
         self.logger.info("Start converting StaticVarCompensator.")
@@ -1446,9 +1245,15 @@ class CimConverter:
         stationSuppliesCim16.StationSuppliesCim16(cimConverter=self).convert_station_supplies_cim16()
         # self._convert_station_supplies_cim16()
         # --------- convert generators ---------
-        self._convert_synchronous_machines_cim16()
-        self._convert_asynchronous_machines_cim16()
-        self._convert_energy_sources_cim16()
+        from .converter_classes.generators import synchronousMachinesCim16
+        synchronousMachinesCim16.SynchronousMachinesCim16(cimConverter=self).convert_synchronous_machines_cim16()
+        # self._convert_synchronous_machines_cim16()
+        from .converter_classes.generators import asynchronousMachinesCim16
+        asynchronousMachinesCim16.AsynchronousMachinesCim16(cimConverter=self).convert_asynchronous_machines_cim16()
+        # self._convert_asynchronous_machines_cim16()
+        from .converter_classes.generators import energySourcesCim16
+        energySourcesCim16.EnergySourceCim16(cimConverter=self).convert_energy_sources_cim16()
+        # self._convert_energy_sources_cim16()
         # --------- convert shunt elements ---------
         self._convert_linear_shunt_compensator_cim16()
         self._convert_nonlinear_shunt_compensator_cim16()
